Diabetes detection code/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Deep Learning & ML
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
Pillow==10.1.0

# Data Processing
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Model Interpretability
tf-keras-vis==0.8.5
lime==0.2.0.1

# Metrics & Evaluation
scikit-image==0.22.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
imbalanced-learn==0.11.0

# Image Augmentation
albumentations==1.3.1
imgaug==0.4.0

# Additional
h5py==3.10.0
joblib==1.3.2
